@@ -1,59 +1,64 @@
 {
   "name": "oerc-client",
-  "version": "1.0.0",
+  "version": "0.9.12",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build-prod": "ng build --configuration production --base-href /client/ --deploy-url /client/ --optimization --build-optimizer --delete-output-path",
+    "build-prod": "ng build --configuration production --base-href /client/ --deploy-url /client/ --delete-output-path",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.5",
-    "@angular/common": "^12.0.5",
-    "@angular/compiler": "^12.0.5",
-    "@angular/core": "^12.0.5",
-    "@angular/forms": "^12.0.5",
-    "@angular/platform-browser": "^12.0.5",
-    "@angular/platform-browser-dynamic": "^12.0.5",
-    "@angular/router": "^12.0.5",
+    "@angular/animations": "^12.2.12",
+    "@angular/common": "^12.2.12",
+    "@angular/compiler": "^12.2.12",
+    "@angular/core": "^12.2.12",
+    "@angular/forms": "^12.2.12",
+    "@angular/localize": "^12.2.12",
+    "@angular/platform-browser": "^12.2.12",
+    "@angular/platform-browser-dynamic": "^12.2.12",
+    "@angular/router": "^12.2.12",
     "@egjs/hammerjs": "^2.0.17",
     "angular-page-visibility": "^11.0.0",
+    "flatpickr": "^4.6.9",
     "keycharm": "^0.4.0",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.33",
     "propagating-hammerjs": "^2.0.1",
     "rxjs": "^6.6.7",
     "spectre.css": "^0.5.9",
-    "tslib": "^2.3.0",
+    "tslib": "^2.3.1",
     "uuid": "^7.0.3",
     "vis-data": "^7.1.2",
-    "vis-timeline": "^7.4.9",
+    "vis-timeline": "^7.5.0",
     "vis-util": "^5.0.2",
     "xss": "^1.0.9",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.0.5",
-    "@angular/cli": "^12.0.5",
-    "@angular/compiler-cli": "^12.0.5",
-    "@types/jasmine": "^3.7.7",
-    "@types/node": "^12.20.15",
+    "@angular-devkit/build-angular": "^12.2.12",
+    "@angular/cli": "^12.2.12",
+    "@angular/compiler-cli": "^12.2.12",
+    "@types/jasmine": "^3.10.1",
+    "@types/node": "^12.20.36",
     "codelyzer": "^6.0.2",
-    "jasmine-core": "~3.6.0",
+    "jasmine-core": "^3.10.1",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "^6.3.4",
+    "karma": "^6.3.6",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.6.0",
-    "postcss": "^8.3.5",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "postcss": "^8.3.11",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
     "typescript": "~4.2.4"
+  },
+  "optionalDependencies": {
+    "fsevents": "^2.3.2"
   }
 }
