@@ -24,16 +24,19 @@ import (
 	"github.com/urfave/cli/v2"
 	"io/ioutil"
 	"log"
+	"net/http"
 	"os"
-	"runtime"
-	"runtime/pprof"
 	"sort"
 	"strconv"
 	"time"
+
+	_ "net/http/pprof"
+
+	"github.com/pkg/profile"
 )
 
 var (
-	version       = "1.0.0"
+	version       = "0.9.12"
 	appConf       AppConfig
 	status        Status
 	verboseGlobal = false
@@ -103,7 +106,9 @@ func main() {
 						log.Printf("Parsing SRF End\n")
 					}
 
-					// update counters
+					FindOverlaps()
+
+					// update counters if something has happened
 					if status.TotalCreatedPE > 0 || status.TotalCreatedTVS > 0 ||
 						status.TotalUpdatedPE > 0 || status.TotalUpdatedTVS > 0 {
 						setSetting(settingKeyLastFetch, time.Now().Format(time.RFC3339))
@@ -151,7 +156,7 @@ func main() {
 					startTime := time.Now()
 					Startup(c)
 					defer Shutdown()
-					log.Printf("Trying to start server at %s:%d ...\n", appConf.ServerHost, appConf.ServerPort)
+					log.Printf("Trying to start server at http://%s:%d ...\n", appConf.ServerHost, appConf.ServerPort)
 
 					defer func() {
 						log.Println("Server end")
@@ -181,6 +186,7 @@ func main() {
 				Aliases: []string{"sc"},
 				Usage:   "Search program data and create recommendations",
 				Action: func(c *cli.Context) error {
+					log.Printf("Starting search process...")
 					startTime := time.Now()
 					Startup(c)
 					defer Shutdown()
@@ -209,7 +215,7 @@ func main() {
 							if c.Bool("force") {
 								ClearLogs()
 							} else {
-								log.Printf("Please set the '--force true' flag to clean the logs from database.\n")
+								log.Printf("Please set the '--force true' flag to clear the logs from database.\n")
 							}
 
 							return nil
@@ -226,7 +232,7 @@ func main() {
 							if c.Bool("force") {
 								ClearRecommendations()
 							} else {
-								log.Printf("Please set the '--force true' flag to clean ALL recommendations from database.\n")
+								log.Printf("Please set the '--force true' flag to clear ALL recommendations from database.\n")
 							}
 
 							return nil
@@ -243,7 +249,24 @@ func main() {
 							if c.Bool("force") {
 								ClearOldRecommendations()
 							} else {
-								log.Printf("Please set the '--force true' flag to clean old recommendations from database.\n")
+								log.Printf("Please set the '--force true' flag to clear old recommendations from database.\n")
+							}
+
+							return nil
+						},
+					},
+					{
+						Name:  "overlaps",
+						Usage: "Clearing overlap status of all program items",
+						Action: func(c *cli.Context) error {
+							log.Println("clear overlaps")
+							Startup(c)
+							defer Shutdown()
+
+							if c.Bool("force") {
+								ClearDeprecations()
+							} else {
+								log.Printf("Please set the '--force true' flag to clear program entries' overlap status from database.\n")
 							}
 
 							return nil
@@ -257,16 +280,32 @@ func main() {
 					},
 				},
 				Action: func(c *cli.Context) error {
-					log.Printf("Clear\n")
+					log.Printf("Clearing all data...\n")
 					Startup(c)
 					defer Shutdown()
 
 					if c.Bool("force") {
+						log.Printf("This could take a while...\n")
 						ClearAll()
 					} else {
 						log.Printf("Please set the '--force true' flag to confirm cleaning the WHOLE database.\n")
 					}
 
+					return nil
+				},
+			},
+			{
+				Name:  "full-overlap-check",
+				Usage: "Run overlap check on all program entries. Could take very long.",
+				Action: func(context *cli.Context) error {
+					startTime := time.Now()
+					Startup(context)
+					defer Shutdown()
+
+					FindOverlapsGlobal()
+
+					duration := time.Now().Sub(startTime)
+					log.Printf("Duration: %.2f Seconds, %.2f Minutes\n", duration.Seconds(), duration.Minutes())
 					return nil
 				},
 			},
@@ -300,34 +339,18 @@ func Startup(c *cli.Context) {
 		verboseGlobal = true
 	}
 
-	// do profiling specific stuff
+	// profiling related stuff
 	if isProfilingEnabled() {
-		dateStr := time.Now().Format(time.RFC3339)
-
-		f, err := ioutil.TempFile("", fmt.Sprintf("oerc-profiling-cpu-%s-*.pprof", dateStr))
-		if err != nil {
-			log.Fatal("could not create CPU profile: ", err)
-		}
-		log.Printf("Profiling (CPU) output is stored in %s\n", f.Name())
-
-		if err := pprof.StartCPUProfile(f); err != nil {
-			log.Fatal("could not start CPU profile: ", err)
-		}
-		shutdownCb = func() {
-			defer f.Close()
-			defer pprof.StopCPUProfile()
-
-			memF, err := ioutil.TempFile("", fmt.Sprintf("oerc-profiling-mem-%s-*.pprof", dateStr))
+		go func() {
+			err := http.ListenAndServe("127.0.0.1:9999", nil)
 			if err != nil {
-				log.Fatal("could not create memory profile: ", err)
+				log.Fatalf("Could not start profiling endpoint: '%v'\n", err)
 			}
-			log.Printf("Profiling (Memory) output is stored in %s\n", memF.Name())
+			log.Printf("Profiling endpoint started at 127.0.0.1:9999")
+		}()
 
-			runtime.GC() // get up-to-date statistics
-			if err := pprof.WriteHeapProfile(memF); err != nil {
-				log.Fatal("could not write memory profile: ", err)
-			}
-			defer memF.Close()
+		shutdownCb = func() {
+			defer profile.Start(profile.MemProfile).Stop()
 		}
 	}
 }
@@ -341,7 +364,7 @@ func Shutdown() {
 		shutdownCb()
 	}
 	// close db
-	db, _ := getDb()
+	db := getDb()
 	s, err := db.DB()
 	if err != nil {
 		log.Fatal(err)
